Moved TextureEditorDesktopApp class and all necessary import statemetents in own file under /src folder
